Introduced lazy proxy around DocRepository
To shift the set-up timing without adding a lot of ugly code (as I did previously), I introduced a lazy dynamic proxy around the repository.

I'm not really happy with this, but I consider it better than the prior approach.  There is also probably a more "groovy" way to do this.

As before, this change is itself non-functional.  The subsequent test will enable spring-data instrumentation and alter the test.