Rename dd-* threads as opentelemetry-* threads.
I grepped for dd-, ignoring add-, comment lines, and some matching guids, and found some threads that needed renaming.
As threads are visible to users, I opted against just dropping the dd- prefix to ensure that it's immediately
obvious that we created the threads when using profilers, thread dumps, etc.