Enable testLatestDeps in PRs (#6744)
but do not make it a required check.

The reason for my change of mind here is:

When new library versions are released which break testLatestDeps, the
quick fix is to temporarily limit the upper bound, but it's not clear
what the quick fix is if we end up merging in a change which breaks
testLatestDeps (short of reverting the change, which then potentially
requires reengaging the contributor after they have moved on from the
PR).