Add limit on ClassLoader WeakReference cache
## Problem
Having many WeakReferences can cause problems in an application. To reduce the number, we have a cache lookup to allow us to reuse those WeakReference instances. This cache is currently unlimited in size, and only reduced by cleaning references to GC'd ClassLoaders.  Most applications have a limited number of persistent ClassLoaders. Some applications generate them dynamically.  This is where the cache can grow excessively.

## Solution
Add a size limit to the cache so that less commonly used loaders are evicted.

Note: This could cause a common entry to be evicted which may result in a less performant lookup (because identity comparison won't match for subsequently generated WeakReferences). This is considered an acceptable tradeoff to avoid excessive memory use.