Dubbo: don't create spans for calls inside the same jvm (#7761)
Resolves
https://github.com/open-telemetry/opentelemetry-java-instrumentation/issues/7520
Invoking a dubbo service, like in
https://github.com/zewade/opentelemetry-dubbo-demo/blob/d7e2417ae787808dbfb43747d586b91a1b29ecec/demo-business-provider/src/main/java/cn/zewade/business/controller/BusinessController.java#L20,
creates a client span (and a server span on the receiving end). If the
caller and service are on the jvm this call doesn't go through remoting
and will be handled with regular java calls. Now when the initially
called service calls another service, like in
https://github.com/zewade/opentelemetry-dubbo-demo/blob/main/demo-business-provider/src/main/java/cn/zewade/business/dubbo/BusinessDubboServiceImpl.java#L25,
that second call will also attempt to create a client span. This second
client span will be suppressed which will also suppress context
propagation.