Fix gradle warning (#7956)
> Java toolchain auto-provisioning enabled, but no java toolchain
repositories declared by the build. Will rely on the built-in
repository. This behaviour has been deprecated and is scheduled to be
removed in Gradle 8.0. In order to declare a repository for java
toolchains, you must edit your settings script and add one via the
toolchainManagement block. See
https://docs.gradle.org/7.6.1/userguide/toolchains.html#sec:provisioning
for more details.

The sample at linked page mentions `repository("adoptium")` which looks
like what we would like to use, but apparently this is just a sample how
configuration might look like, `AdoptiumResolver` doesn't really exist.
The one that exists is https://github.com/gradle/foojay-toolchains