Standardize on “Javaagent” instead of “java agent”
There are way more usages of “javaagent” within the project, so the inverse would be much more challenging.

I find the inconsistency confusing. The JVM calls it a `javaagent` so we should also.

Some minor word rewrapping snuck in because of my IDE settings.  If this is problematic, I can roll that back.