Resolve `ParameterNameDiscoverer` Bean Conflict in `spring-boot-autoconfigure`
This update addresses a conflict arising in the `spring-boot-autoconfigure`
project when using `opentelemetry-spring-boot-starter` and other projects
which use the `ParameterNameDiscoverer` bean simultaneously, e.g.
`springdoc-openapi-starter-webmvc-ui`.
The issue occurs due to the presence of multiple candidates for the
`ParameterNameDiscoverer` object, leading to an injection conflict.

When both starters are used together, the application fails to start,
displaying an error related to the ParameterNameDiscoverer object.
This is due to the ambiguity in selecting the correct bean to inject,
as multiple candidates are present.

Example of the error output:

```
***************************
APPLICATION FAILED TO START
***************************

Description:

Parameter 1 of method otelInstrumentationWithSpanAspect in io.opentelemetry.instrumentation.spring.autoconfigure.instrumentation.annotations.InstrumentationAnnotationsAutoConfiguration required a single bean, but 2 were found:
        - parameterNameDiscoverer: defined by method 'parameterNameDiscoverer' in class path resource [io/opentelemetry/instrumentation/spring/autoconfigure/instrumentation/annotations/InstrumentationAnnotationsAutoConfiguration.class]
        - localSpringDocParameterNameDiscoverer: defined by method 'localSpringDocParameterNameDiscoverer' in class path resource [org/springdoc/core/configuration/SpringDocConfiguration.class]

Action:

Consider marking one of the beans as @Primary, updating the consumer to accept multiple beans, or using @Qualifier to identify the bean that should be consumed
```

To resolve this issue, this commit defines the private final `ParameterNameDiscoverer` instance
and use that directly instead of injection in `InstrumentationAnnotationsAutoConfiguration`.

Signed-off-by: moznion <moznion@mail.moznion.net>