Implements getList() in the Spring Web MVC instrumentation's JakartaHttpServletRequestGetter.
Implements a test to demonstrate the new W3CBaggagePropagator working with multiple baggage headers (using a locally published version of opentelemetry-api and opentelemetry-context).

The TextMapGetter operates on a Jakarta servlet HttpServletRequest, a very common interface used in many servers. This demonstrates that using getList() will work in many area / servers.