Library instrumentation for Java Servlet 3.0 Filters.
Includes handling for attaching all the async stuff, but can't support everything the java agent does.

I could use help figuring out the right way to shadow all the necessary dependencies.  I'm a bit rusty with our shadow usage.
Depending on how aggressive we want to be, the dependency could be reversed so that `:servlet-3.0:javaagent` depends on this new module, but that's more aggressive.  Perhaps shadow could prune unused classes (with `minimize()`) instead.

Once we get the modules figured out, I can work on adding some unit tests.

If we like this approach, it could probably be copied and modified pretty easily to support servlet 5, but I'm not familiar with that instrumentation.